Fix data e inserir novo servico

+++ src/main/java/com/example/ordermServico/repositories/ServicoPrestadoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ordermServico.entities.ServicoPrestado;
 
-public interface ServicoPrestadoRepository extends JpaRepository<ServicoPrestado, Long>{
+public interface ServicoPrestadoRepository extends JpaRepository<ServicoPrestado, Integer>{
 }
